Reintroduced separate classes for items with special renderers
This is due to sidedness & classloading. Client classes do not exist on the dedicated server. May be more inline with the design philosophy of this system as well, since the reason they have the initializeClient method in the item in the first place is so that other mods can't swoop in and change the renderer without the original mod author knowing.